requirements.txt
google-api-python-client==2.118.0
google-auth==2.23.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
requests==2.31.0
urllib3==1.26.15
